# Conflicts:
#	compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/BIRPackageSymbolEnter.java
#	compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/bir/BIRGen.java
#	compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/desugar/ClosureGenerator.java
#	compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/desugar/Desugar.java
#	compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/semantics/model/symbols/BObjectTypeSymbol.java
#	compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/semantics/model/symbols/BRecordTypeSymbol.java

+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/semantics/model/symbols/BRecordTypeSymbol.java
@@ -34,7 +34,7 @@ import java.util.Map;
 public class BRecordTypeSymbol extends BStructureTypeSymbol {
 
     public Map<String, BInvokableSymbol> defaultValues;
-    public BRecordTypeSymbol(int symTag, long flags, Name name, PackageID pkgID, BType type, BSymbol owner,
+    public BRecordTypeSymbol(long symTag, long flags, Name name, PackageID pkgID, BType type, BSymbol owner,
                              Location pos, SymbolOrigin origin) {
         super(SymbolKind.RECORD, symTag, flags, name, pkgID, type, owner, pos, origin);
         this.defaultValues = new HashMap<>();
